Move Midi classes to their own folder, clearer folder boundaries
