+++ JammaLib/src/graphics/LoopModel.cpp
