+++ JammaLib/src/graphics/LoopModel.h
@@ -5,7 +5,7 @@
 #include <vector>
 #include <tuple>
 #include <memory>
-#include "../include/Constants.h"
+#include "../../include/Constants.h"
 #include "../utils/ArrayUtils.h"
 #include "../utils/VecUtils.h"
 #include "../audio/BufferBank.h"
