+++ JammaLib/src/graphics/QuantisationModel.h
@@ -2,7 +2,7 @@
 
 #include <vector>
 #include "../gui/GuiModel.h"
-#include "Timer.h"
+#include "../engine/Timer.h"
 
 namespace engine
 {
