+++ JammaLib/src/midi/MidiEvent.h
