+++ JammaLib/src/midi/MidiLoop.cpp
