+++ JammaLib/src/midi/MidiNoteSpan.cpp
