+++ JammaLib/src/midi/MidiNoteSpan.h
